Rename model manager files
